Function ra_prof::profile

pub fn profile(label: &'static str) -> Profiler

This function starts a profiling scope in the current execution stack with a given description. It returns a Profile structure and measure elapsed time between this method invocation and Profile structure drop. It supports nested profiling scopes in case when this function invoked multiple times at the execution stack. In this case the profiling information will be nested at the output. Profiling information is being printed in the stderr.

Example

use ra_prof::{profile, set_filter, Filter};

let f = Filter::from_spec("profile1|profile2@2");
set_filter(f);
profiling_function1();

fn profiling_function1() {
    let _p = profile("profile1");
    profiling_function2();
}

fn profiling_function2() {
    let _p = profile("profile2");
}

This will print in the stderr the following:

 0ms - profile
     0ms - profile2