Function ra_prof::set_filter

pub fn set_filter(f: Filter)

Set profiling filter. It specifies descriptions allowed to profile. This is helpful when call stack has too many nested profiling scopes. Additionally filter can specify maximum depth of profiling scopes nesting.

#Example

use ra_prof::{set_filter, Filter};
let f = Filter::from_spec("profile1|profile2@2");
set_filter(f);